+++ src/fft/mod.rs
@@ -1,6 +1,7 @@
 mod butterflies;
 pub(crate) mod cooley_tukey_radix2;
 pub(crate) mod cooley_tukey_radixn;
+pub mod planner;
 mod radix_fft;
 
 pub(crate) use radix_fft::{Forward, Inverse, Radix, RadixFFT};